+++ src/modules/adapter.js
