+++ src/modules/buffer.js
