--- src/modules/ui-scroll-viewport.js
@@ -1,12 +1,0 @@
-export default function uiScrollViewport() {
-	return {
-		controller: [
-			'$scope',
-			'$element',
-			function (scope, element) {
-				this.viewport = element;
-				return this;
-			}
-		]
-	};
-}
